refactor(error-handlers): BUG-002 proper fix + error_handlers.py split + DAG/DRY/security fixes (Issue #122)
- BUG-002: InvalidRoleError 専用例外を新設し _parse_role() で ValueError を変換
  グローバル ValueError ハンドラを廃止（スコープ限定の専用 handler に切り替え）
- error_handlers.py (607行) を error_handlers/ パッケージに責務分割
  (_common / core / empire / task / deliverable_template / middleware / __init__)
- _clean_domain_message() / error_response() を _common.py に集約（DRY）
- CompositionCycleError レスポンスから cycle_path を除外（最小情報開示原則 / OWASP A05/A09）
- _check_dag を BFS から DFS + ancestors frozenset に置換（diamond-DAG 偽陽性バグ修正）
- update() 内のデッドコード削除（create_new_version の未使用代入）
