+++ pyproject.toml
@@ -12,6 +12,7 @@ dev = [
     "pytest-asyncio>=0.24",
     "pytest-cov>=7.1.0",
     "httpx>=0.25",
+    "ruff>=0.15.12",
 ]
 
 [tool.uv.sources]
